refactor(slots): extract metrics/cgroup/systemd IO adapters to slots/metrics_collect
P3-routers §J: move the route layer's heaviest non-policy code out of
routes/slots.py into a new deep module hal0.slots.metrics_collect — the
systemctl-show adapter, the cgroup-v2 memory.current walk, the llama-server
/metrics+/slots scrape, the per-slot fan-out, and the rolling tok/s + TTFT
window views. slot_metrics (GET /metrics) is now a request→service→envelope
shell. Route paths/methods/payloads/status codes unchanged.

Back-compat: routes/slots keeps underscore-named re-export bindings
(_scrape_llama_metrics, _systemd_show, _docker_container_mem_bytes,
_tps_from_events) plus thin request adapters (_per_slot_local_tps,
_per_slot_ttft, _local_slot_metrics) so hal0.metrics.sampler's import of
_scrape_llama_metrics and the test-suite's patch/import seams keep working
without edits.

slots.py: 1964 → 1629 lines.
